% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcGSAtest.R
\docType{data}
\name{dcGSAtest}
\alias{dcGSAtest}
\title{dcGSA test data}
\description{
A R data object of example data to test dcGSA. This is a list comprised of
ID, data (phenotypes of interest), gene (longitudinal gene expresion profiles).
}
\examples{
data(dcGSAtest) # load the test dataset
}
\keyword{data}

